# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"desafio-bootcamp","description":"pens-luis-bootcamp-final","contact":{"name":"pens-luis-bootcamp-final","email":"luis.pens@mercadolibre.com"},"version":"last"},"servers":[{"url":"https://lalala-integration_test_desafio-bootcamp.furyapps.io","description":"Scope lalala-integration_test"}],"paths":{"/api/v1/users/new":{"post":{"tags":["users-controller"],"operationId":"createUser","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts":{"post":{"tags":["parts-controller"],"operationId":"createPart","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewPartDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"patch":{"tags":["parts-controller"],"operationId":"createPart_1","parameters":[{"name":"partCode","in":"query","required":true,"schema":{"type":"string"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/price":{"post":{"tags":["parts-controller"],"operationId":"newPrice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartRecordDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/orders":{"get":{"tags":["orders-controller"],"operationId":"getOrders","parameters":[{"name":"dealerNumber","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"deliveryStatus","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponseDto"}}}}}},"post":{"tags":["orders-controller"],"operationId":"generateOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenerateOrderResponseDto"}}}}}}},"/api/v1/users/changeRole":{"patch":{"tags":["users-controller"],"operationId":"changeRole","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/orders/update_status":{"patch":{"tags":["orders-controller"],"operationId":"udpadteOrderStatus","parameters":[{"name":"orderStatus","in":"query","required":true,"schema":{"type":"string"}},{"name":"orderNumberCM","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BasicResponseDto"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/pricehistory/{partCode}":{"get":{"tags":["parts-controller"],"operationId":"getPriceHistory","parameters":[{"name":"partCode","in":"path","required":true,"schema":{"type":"string"}},{"name":"fromDate","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/orders/{orderId}":{"get":{"tags":["orders-controller"],"operationId":"getOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderStatusDto"}}}}}}},"/api/v1/parts/list":{"get":{"tags":["parts-controller"],"operationId":"getParts","parameters":[{"name":"queryType","in":"query","required":false,"schema":{"type":"string","default":"C"}},{"name":"date","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartResponseDto"}}}}}}}},"components":{"schemas":{"NewPartDto":{"type":"object","properties":{"description":{"type":"string"},"stock":{"type":"integer","format":"int32"},"partCode":{"type":"string"},"providerId":{"type":"integer","format":"int64"},"netWeight":{"type":"integer","format":"int32"},"longDimension":{"type":"integer","format":"int32"},"widthDimension":{"type":"integer","format":"int32"},"talDimension":{"type":"integer","format":"int32"},"lastModification":{"type":"string","format":"date"},"discountTypeId":{"type":"integer","format":"int64"},"normalPrice":{"type":"number","format":"double"},"urgentPrice":{"type":"number","format":"double"}}},"PartRecordDto":{"type":"object","properties":{"discountType":{"type":"integer","format":"int64"},"normalPrice":{"type":"number","format":"double"},"urgentPrice":{"type":"number","format":"double"},"partCode":{"type":"string"}}},"OrderDetailDto":{"type":"object","properties":{"partCode":{"type":"string"},"description":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"accountType":{"type":"string"},"reason":{"type":"string"}}},"OrderRequestDto":{"type":"object","properties":{"centralHouseId":{"type":"integer","format":"int64"},"consessionarieId":{"type":"integer","format":"int64"},"shippingType":{"type":"string"},"parts":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDto"}}}},"GenerateOrderResponseDto":{"type":"object","properties":{"orderCMId":{"type":"string"},"centralHouseId":{"type":"integer","format":"int64"}}},"BasicResponseDto":{"type":"object","properties":{"httpStatus":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"message":{"type":"string"}}},"OrderDto":{"type":"object","properties":{"orderNumber":{"type":"integer","format":"int32"},"orderDate":{"type":"string"},"deliveryDate":{"type":"string"},"daysDelayed":{"type":"integer","format":"int32"},"deliveryStatus":{"type":"string"},"orderDetails":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDto"}}}},"OrderResponseDto":{"type":"object","properties":{"orders":{"type":"array","items":{"$ref":"#/components/schemas/OrderDto"}},"dealerNumber":{"type":"integer","format":"int64"}}},"OrderStatusDto":{"type":"object","properties":{"orderNumberCE":{"type":"string"},"orderDate":{"type":"string"},"orderStatus":{"type":"string"},"orderDetails":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDto"}}}},"PartDto":{"type":"object","properties":{"description":{"type":"string"},"marker":{"type":"string"},"quantity":{"type":"string"},"partCode":{"type":"string"},"discountType":{"type":"string"},"normalPrice":{"type":"string"},"urgentPrice":{"type":"string"},"netWeight":{"type":"string"},"longDimension":{"type":"string"},"widthDimension":{"type":"string"},"tallDimension":{"type":"string"},"lastModification":{"type":"string"}}},"PartResponseDto":{"type":"object","properties":{"parts":{"type":"array","items":{"$ref":"#/components/schemas/PartDto"}}}}}}}